+++ flutter_extensions/lib/src/regexp.dart
@@ -1,5 +1,5 @@
 extension RegExpMatchExt on RegExpMatch {
-  String tryNamedGroup(String name) {
+  String? tryNamedGroup(String name) {
     try {
       return namedGroup(name);
     } catch (_) {