+++ pure_extensions/lib/src/dart/primitives/string.dart
@@ -13,7 +13,7 @@ extension StringDartExtensions on String {
   /// hello_word_good => helloWordGood
   String toCamelCase() {
     return replaceAllMapped(RegExp(r'([_:].)'), (match) {
-      return match.group(0)[1].toUpperCase();
+      return match.group(0)![1].toUpperCase();
     });
   }
 
@@ -23,13 +23,13 @@ extension StringDartExtensions on String {
 
   String ifBlank(String Function() fn) => isBlank ? fn() : this;
 
-  String nullIfEmpty() => isEmpty ? null : this;
+  String? nullIfEmpty() => isEmpty ? null : this;
 
-  Map<String, dynamic> decodeToMap() => jsonDecode(this) as Map<String, dynamic>;
+  Map<String, dynamic>? decodeToMap() => jsonDecode(this) as Map<String, dynamic>?;
 
-  List<Map<String, dynamic>> decodeToList() => jsonDecode(this) as List<Map<String, dynamic>>;
+  List<Map<String, dynamic>>? decodeToList() => jsonDecode(this) as List<Map<String, dynamic>>?;
 
-  List<String> divide([int startAt = 0, int divideAt, int endAt]) {
+  List<String> divide([int startAt = 0, int? divideAt, int? endAt]) {
     endAt ??= length;
     divideAt ??= endAt ~/ 2;
     return [
@@ -38,7 +38,7 @@ extension StringDartExtensions on String {
     ];
   }
 
-  String advSubString(int start, [int end]) {
+  String advSubString(int start, [int? end]) {
     return substring(
       start < 0 ? start + length : start,
       end != null && end < 0 ? length + end : end,
@@ -57,16 +57,16 @@ extension StringDartExtensions on String {
   }
 
   num toNum() => num.parse(this);
-  num toTryNum() => num.tryParse(this);
+  num? toTryNum() => num.tryParse(this);
 
-  int toInt({int radix}) => int.parse(this, radix: radix);
-  int toTryInt({int radix}) => int.tryParse(this, radix: radix);
+  int toInt({int? radix}) => int.parse(this, radix: radix);
+  int? toTryInt({int? radix}) => int.tryParse(this, radix: radix);
 
   double toDouble() => double.parse(this);
-  double toTryDouble() => double.tryParse(this);
+  double? toTryDouble() => double.tryParse(this);
 
   Rational toRational() => Rational.parse(this);
-  Rational tryToRational() {
+  Rational? tryToRational() {
     try {
       return Rational.parse(this);
     } catch (_) {
@@ -74,9 +74,9 @@ extension StringDartExtensions on String {
     }
   }
 
-  Uri toUri([int start = 0, int end]) => Uri.parse(this, start, end);
-  Uri toTryUri([int start = 0, int end]) => Uri.tryParse(this, start, end);
+  Uri toUri([int start = 0, int? end]) => Uri.parse(this, start, end);
+  Uri? toTryUri([int start = 0, int? end]) => Uri.tryParse(this, start, end);
 
   DateTime toDateTime() => DateTime.parse(this);
-  DateTime toTryDateTime() => DateTime.tryParse(this);
+  DateTime? toTryDateTime() => DateTime.tryParse(this);
 }